Revert "1565"
This reverts commit 08b73c7bffb70ab8ae72b3e45aea0f7da499c0a8.

+++ src/playground_base/playground_base/gui/controls/FBOutputParamLabel.hpp
@@ -25,5 +25,5 @@ public:
   juce::String getTooltip() override;
   void parentHierarchyChanged() override;
   int FixedWidth(int height) const override;
-  void SetValueNormalizedFromHost(double normalized) override;
+  void SetValueNormalizedFromHost(float normalized) override;
 };
