+++ src/playground_base/playground_base/gui/controls/FBParamSlider.hpp
@@ -35,5 +35,5 @@ public:
   double getValueFromText(const juce::String& text) override;
 
   void UpdateExchangeState();
-  void SetValueNormalizedFromHost(double normalized) override;
+  void SetValueNormalizedFromHost(float normalized) override;
 };
